Make PaginatedResource a type of GlobusResponse
Now that GlobusResponses can be more variegated and featureful, it makes
sense to make PaginatedResource a type of GlobusResponse which has its
`data` property as a list of its contents.
